+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0] - Unreleased
+
+### Added
+- **Row Visualization**: New built-in row visualization that allows you to layout status indicators side-by-side
+
+### Changed
+- **Logo**: New project logo
+- **Stack Visualization**
+ - Layout groups can now be laid out column-wise
+ - Ordering of status indicators can now be controlled
+- **SNMP Monitor**:
+ - Retry improvements
+
 ## [0.15.1] - 2025-09-04
 
 ### Changed